Move checked_mutex.hpp out of object-store
