+++ Package.swift
@@ -493,7 +493,6 @@ let headers: [String] = [
     "realm/object-store/util/bson/min_key.hpp",
     "realm/object-store/util/bson/mongo_timestamp.hpp",
     "realm/object-store/util/bson/regular_expression.hpp",
-    "realm/object-store/util/checked_mutex.hpp",
     "realm/object-store/util/copyable_atomic.hpp",
     "realm/object-store/util/event_loop_dispatcher.hpp",
     "realm/object-store/util/generic/scheduler.hpp",
@@ -588,6 +587,7 @@ let headers: [String] = [
     "realm/util/call_with_tuple.hpp",
     "realm/util/cf_ptr.hpp",
     "realm/util/cf_str.hpp",
+    "realm/util/checked_mutex.hpp",
     "realm/util/circular_buffer.hpp",
     "realm/util/cli_args.hpp",
     "realm/util/copy_dir_recursive.hpp",
