+++ src/realm/CMakeLists.txt
@@ -218,6 +218,7 @@ set(REALM_INSTALL_HEADERS
     util/buffer_stream.hpp
     util/call_with_tuple.hpp
     util/cf_ptr.hpp
+    util/checked_mutex.hpp
     util/encrypted_file_mapping.hpp
     util/errno.hpp
     util/features.h
