+++ src/realm/object-store/CMakeLists.txt
@@ -76,7 +76,6 @@ set(HEADERS
 
     util/aligned_union.hpp
     util/atomic_shared_ptr.hpp
-    util/checked_mutex.hpp
     util/copyable_atomic.hpp
     util/event_loop_dispatcher.hpp
     util/scheduler.hpp
