+++ src/realm/object-store/impl/realm_coordinator.hpp
@@ -21,8 +21,7 @@
 
 #include <realm/object-store/shared_realm.hpp>
 
-#include <realm/object-store/util/checked_mutex.hpp>
-
+#include <realm/util/checked_mutex.hpp>
 #include <realm/version_id.hpp>
 
 #include <condition_variable>
